move RunRouteControllerManager to a separate package
