+++ pkg/cmd/route-controller-manager/controller_manager.go
@@ -8,8 +8,8 @@ import (
 	openshiftcontrolplanev1 "github.com/openshift/api/openshiftcontrolplane/v1"
 	"github.com/openshift/library-go/pkg/serviceability"
 
-	origincontrollers "github.com/openshift/openshift-controller-manager/pkg/cmd/controller"
 	routecontrollers "github.com/openshift/openshift-controller-manager/pkg/cmd/controller/route"
+	origincontrollers "github.com/openshift/openshift-controller-manager/pkg/cmd/routecontroller"
 	"github.com/openshift/openshift-controller-manager/pkg/routeversion"
 )
