+++ pkg/cmd/routecontroller/route.go
@@ -1,4 +1,4 @@
-package controller
+package routecontroller
 
 import (
 	"context"
